refactor: AuthConstant 패키지 변경 (#230)
